Cumulative fixes based on GekkoCore unit test results

+++ SRC/GekkoCore/Gekko.cpp
@@ -65,25 +65,25 @@ namespace Gekko
 
         // BAT registers are scattered across the SPR address space. This is not very convenient, we will make it convenient.
 
-        dbatu[0] = &regs.spr[(int)SPR::DBAT0U];
-        dbatu[1] = &regs.spr[(int)SPR::DBAT1U];
-        dbatu[2] = &regs.spr[(int)SPR::DBAT2U];
-        dbatu[3] = &regs.spr[(int)SPR::DBAT3U];
+        dbatu[0] = &regs.spr[SPR::DBAT0U];
+        dbatu[1] = &regs.spr[SPR::DBAT1U];
+        dbatu[2] = &regs.spr[SPR::DBAT2U];
+        dbatu[3] = &regs.spr[SPR::DBAT3U];
 
-        dbatl[0] = &regs.spr[(int)SPR::DBAT0L];
-        dbatl[1] = &regs.spr[(int)SPR::DBAT1L];
-        dbatl[2] = &regs.spr[(int)SPR::DBAT2L];
-        dbatl[3] = &regs.spr[(int)SPR::DBAT3L];
+        dbatl[0] = &regs.spr[SPR::DBAT0L];
+        dbatl[1] = &regs.spr[SPR::DBAT1L];
+        dbatl[2] = &regs.spr[SPR::DBAT2L];
+        dbatl[3] = &regs.spr[SPR::DBAT3L];
 
-        ibatu[0] = &regs.spr[(int)SPR::IBAT0U];
-        ibatu[1] = &regs.spr[(int)SPR::IBAT1U];
-        ibatu[2] = &regs.spr[(int)SPR::IBAT2U];
-        ibatu[3] = &regs.spr[(int)SPR::IBAT3U];
+        ibatu[0] = &regs.spr[SPR::IBAT0U];
+        ibatu[1] = &regs.spr[SPR::IBAT1U];
+        ibatu[2] = &regs.spr[SPR::IBAT2U];
+        ibatu[3] = &regs.spr[SPR::IBAT3U];
 
-        ibatl[0] = &regs.spr[(int)SPR::IBAT0L];
-        ibatl[1] = &regs.spr[(int)SPR::IBAT1L];
-        ibatl[2] = &regs.spr[(int)SPR::IBAT2L];
-        ibatl[3] = &regs.spr[(int)SPR::IBAT3L];
+        ibatl[0] = &regs.spr[SPR::IBAT0L];
+        ibatl[1] = &regs.spr[SPR::IBAT1L];
+        ibatl[2] = &regs.spr[SPR::IBAT2L];
+        ibatl[3] = &regs.spr[SPR::IBAT3L];
 
         // Registers
 
@@ -93,9 +93,9 @@ namespace Gekko
         regs.msr &= ~(MSR_DR | MSR_IR);
 
         regs.tb.uval = 0;
-        regs.spr[(int)SPR::HID1] = 0x8000'0000;
-        regs.spr[(int)SPR::DEC] = 0;
-        regs.spr[(int)SPR::CTR] = 0;
+        regs.spr[SPR::HID1] = 0x8000'0000;
+        regs.spr[SPR::DEC] = 0;
+        regs.spr[SPR::CTR] = 0;
 
         gatherBuffer->Reset();
 
@@ -113,9 +113,9 @@ namespace Gekko
     {
         regs.tb.uval += CounterStep;         // timer
 
-        uint32_t old = regs.spr[(int)SPR::DEC];
-        regs.spr[(int)SPR::DEC]--;          // decrementer
-        if ((old ^ regs.spr[(int)SPR::DEC]) & 0x80000000)
+        uint32_t old = regs.spr[SPR::DEC];
+        regs.spr[SPR::DEC]--;          // decrementer
+        if ((old ^ regs.spr[SPR::DEC]) & 0x80000000)
         {
             if (regs.msr & MSR_EE)
             {
@@ -186,26 +186,26 @@ namespace Gekko
 
         // save regs
 
-        regs.spr[(int)Gekko::SPR::SRR0] = regs.pc;
-        regs.spr[(int)Gekko::SPR::SRR1] = regs.msr;
+        regs.spr[Gekko::SPR::SRR0] = regs.pc;
+        regs.spr[Gekko::SPR::SRR1] = regs.msr;
 
         // Special processing for MMU
         if (code == Exception::ISI)
         {
-            regs.spr[(int)Gekko::SPR::SRR1] &= 0x0fff'ffff;
+            regs.spr[Gekko::SPR::SRR1] &= 0x0fff'ffff;
 
             switch (MmuLastResult)
             {
                 case MmuResult::PageFault:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x4000'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x4000'0000;
                     break;
 
                 case MmuResult::ProtectedFetch:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x0800'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x0800'0000;
                     break;
 
                 case MmuResult::NoExecute:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x1000'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x1000'0000;
                     break;
                 
                 default:
@@ -214,20 +214,20 @@ namespace Gekko
         }
         else if (code == Exception::DSI)
         {
-            regs.spr[(int)Gekko::SPR::DSISR] = 0;
+            regs.spr[Gekko::SPR::DSISR] = 0;
 
             switch (MmuLastResult)
             {
                 case MmuResult::PageFault:
-                    regs.spr[(int)Gekko::SPR::DSISR] |= 0x4000'0000;
+                    regs.spr[Gekko::SPR::DSISR] |= 0x4000'0000;
                     break;
 
                 case MmuResult::ProtectedRead:
-                    regs.spr[(int)Gekko::SPR::DSISR] |= 0x0800'0000;
+                    regs.spr[Gekko::SPR::DSISR] |= 0x0800'0000;
                     break;
                 
                 case MmuResult::ProtectedWrite:
-                    regs.spr[(int)Gekko::SPR::DSISR] |= 0x0A00'0000;
+                    regs.spr[Gekko::SPR::DSISR] |= 0x0A00'0000;
                     break;
 
                 default:
@@ -238,21 +238,21 @@ namespace Gekko
         // Special processing for Program
         if (code == Exception::PROGRAM)
         {
-            regs.spr[(int)Gekko::SPR::SRR1] &= 0x0000'ffff;
+            regs.spr[Gekko::SPR::SRR1] &= 0x0000'ffff;
 
             switch (PrCause)
             {
                 case PrivilegedCause::FpuEnabled:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x0010'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x0010'0000;
                     break;
                 case PrivilegedCause::IllegalInstruction:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x0008'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x0008'0000;
                     break;
                 case PrivilegedCause::Privileged:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x0004'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x0004'0000;
                     break;
                 case PrivilegedCause::Trap:
-                    regs.spr[(int)Gekko::SPR::SRR1] |= 0x0002'0000;
+                    regs.spr[Gekko::SPR::SRR1] |= 0x0002'0000;
                     break;
                 default:
                     break;
